Fix validation popup.
Since commit 812e51a0c53abfe70393c56e5b053dd6206e7845 it was not
possible to close the validation popup with the keyboard by pressing
return after selecting the cancel button. Bug was due to the main window
emiting a key release event for the return key after the validation
popup was closed.