Add note categories.
I completly removed the hidden state of notes from the note model.
Now to be hidden, a note has to be in at least one hidden note category.

This commit contains a lot of test fixes, I exceeded the maximum number
of tests we could have without exhausting the number of open files
because we leaked redis connection for each test.