Fix a bug with panels management.
If a category had a product that had the name of another catgeory AND
the latter had an ID greater than the catgeory containing the product
then, the category would appear empty on the left-hand side of the
panels management window.

Since I couldn't find why this condition was here, I just removed it.