Don't deadlock if an imported CSV is not correctly formated.
Currently if the CSV file was malformed, the window was never shown and
thus the closed signal was never sent, resulting in the notes_management
lock never being released.
Raising an exception if the CSV file is malformed and catching it at the
window creation allow us to release the key directly.