Fix the CSV import when trying to import fails.
If the import fails on the first try, then the window is left in a state
where it thinks that the import has been validated and shouldn't allow
another transaction to be issued (this is done this way to avoid
importing a CSV twice by error). Now we reset the validated state when
an import fails.