Cache the synced settings even if their type is dict
While playing with connection loss to redis and the database to fix some
freezes, I saw that the synced setting `ECOCUP_PRICE_TYPES` wasn't
cached and a call to redis was done every time somebody added an ecocup.

This was easily fixed by setting the value variable to the return value
for that setting instead of returning it directly. Synced settings which are
dicts now have the same behavior as every other setting and are correctly
cached.