Update pipeline
This commit updates the CI pipeline. Tests are now run in an isolated
environment using docker. Stages are separated in a semantically correct
way allowing us to know what is running and what is the purpose of the
CI job. Since the environments are isolated, tests can be run in
parallel.

Code have been updated in order to get the database informations from
environment first and fallback to the configuration if no environment
variable were found.

A CI image build can be scheduled on the CI server. This allows to
maintain an up to date test image while using a pre built image,
accelerating the tests. After a build is complete the complete pipeline
is tested to ensure the new image doesn't trigger new tests fail
indicating a lack of compatibility between the dependencies versions and
the software. This approach has some drawbacks:
    - Adding new dependencies requires to build a new CI image or wait
    until the CI creates another one
    - Without using a private docker registry, this method required a
    particular setup: A shell runner with docker access right must run
    on the same machine as the docker runner executing the test
    pipeline. A setup using multiple machine is not possible.