Start the transition to rust for products
This one was a bit more complicated as it needed some bindings to use
BigDecimal. I added a diesel_helpers with a wrapper around
bigdecimal::BigDecimal, this is the only type that should ever be passed
in the application until it's not needed anymore (we can then change our
imports to the real bigdecimal).
The py_helpers should contain implementations to convert a wrapper type to a
PyObject and the other way around. In this case we never need to get
from a python decimal to a BigDecimal so I didn't write it.