Port api.users.get_rights to rust
This is the first getter getting ported to rust and it needed a dict!
macro to avoid creating another Model struct or writing some very boring
code so I added that in the py_helpers module. The rest is pretty
straightforward, get the user and return the rights dicts made with the
user's values.