When an exception is raised, call sys.exit(1)
Qt sucks really hard and when an exception is raised in some
QtCallbackThingy, it just segfaults or, if it has a custom excepthook,
do nothing and we can't detect that the test failed. Now we just call
sys.exit(1) when a test fail which means that it will be detected as
failed on CI but it will be a pain while developping since the test
suite will stop after the first test failure. Thanks a lot PyQt !