Port api.panels.add_product[s] to rust
This one was a bit tricky as the python API is pretty bad in that area
compared to what would look like a good API on the rust side. This
explains why I'm doing a SQL query in the py_rust binding function to
avoid calling SELECT a million times if we're moving lots of products to
a panel.