Change the note column in note_categories_assoc
While generating tables! calls for diesel, I saw that this column didn't
have the right type as it's a FK to notes.id, it should be BIGINT to
match the BIGSERIAL (why is it even a bigserial ???). I chose to keep it
that way in case the production DB has some weird stuff in to avoid
breaking anything.