Cache note stats instead of computing them...
The notes stats (total refilled/consumed) are now stored alongside other
notes informations and are updated on every transaction. This will avoid
computing them at launch time or when a transaction is issued (since
it's pretty slow as it requires a full history scan).

For that to work without adding too much code everywhere, I removed the
`api.notes.transactions` function and put that logic in SQL triggers
instead. This means that the history of a note and the amount on it are
now linked at all times, it's not possible anymore for notes to be
desynchronized with their history.

This is a huge breaking change hence the `VERSION` bump. Any old
instance would otherwise duplicate transactions as they would update the
note amount manually and via the transactions triggers and we don't want
that.