Use lazy_static to import python modules in rapi::py_helpers
This removes our direct usage of unsafe code and is way easier to grasp
and copy/paste to add new imports.
Instead of waiting for the git from the calls, we call
Python::acquire_gil directly in the lazy_static initialization. I don't
know why I didn't think about that earlier. This also remove the usage
of the Box and the option but the benchmarks didn't show any improvement
(nor regression) from that.