Disable the cache during the build as it's creating more troubles than it's solving
Gitlab can't autoclean its own cache every N days meaning that it can
become quite large because of all the rust updates. This slows down
things a lot and leads to timeout.
With a fresh cache, we don't kill our disks and tests seem to be running
just fine in an acceptable time frame.